requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0
python-multipart>=0.0.6
pydantic>=2.0.0
openai>=1.0.0
scanpy>=1.9.0
anndata>=0.9.0
squidpy>=1.2.0
pandas>=2.0.0
numpy>=1.24.0,<2  # pyradiomics 与 NumPy 2.x 不兼容，需 1.x
matplotlib>=3.7.0
streamlit>=1.28.0
networkx>=3.0
pyyaml>=6.0
jinja2>=3.1.0
# Celery 和 Redis 支持
celery>=5.3.0
redis>=5.0.0
flower>=2.0.0
# SSH 远程执行支持
paramiko>=3.0.0
# HTTP 请求库（metabolomics_tool 需要）
requests>=2.31.0
# 科学计算库（metabolomics_tool 需要）
scipy>=1.11.0
scikit-learn>=1.3.0
seaborn>=0.12.0
statsmodels>=0.14.0
# 表格格式化（pandas to_markdown 需要）
tabulate>=0.9.0
# 通路富集分析（metabolomics_pathway_enrichment 需要）
gseapy>=1.0.0
# 细胞类型注释（scanpy_tool 可选功能）
celltypist>=1.6.0
# 双联体检测（必需）
scrublet>=0.2.3
# Tool-RAG 架构：Vector Database 和 Embeddings
langchain-chroma>=0.1.0
langchain-ollama>=0.1.0
langchain-core>=0.1.0
chromadb>=0.4.0


# File signing (BLAKE3 + Ed25519), auto-init keys
blake3>=0.3.3
pynacl>=1.5.0

# 影像组学 (Radiomics) 模态
SimpleITK>=2.3.0
pyradiomics>=3.1.0
opencv-python-headless>=4.8.0

# Phase 1: 用户与资产中台 - MySQL + SQLAlchemy
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=41.0.0

# Phase 2: 鉴权 - 密码哈希与 JWT
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

